small required changes and object b taking decimal values now

+++ src/ObjectB/models.py
@@ -1,11 +1,11 @@
 from django.db import models
 
 class ObjBModel(models.Model):
-    oilPrice = models.IntegerField()
-    oilSD = models.IntegerField()
-    gasPrice = models.IntegerField()
-    gasSD = models.IntegerField()
-    percLine = models.IntegerField()
+    oilPrice = models.FloatField()
+    oilSD = models.FloatField()
+    gasPrice = models.FloatField()
+    gasSD = models.FloatField()
+    percLine = models.FloatField()
 
     def __str__(self):
         return self.oilPrice + ' ' + self.gasPrice
